+++ hooks/hooks.ts
@@ -1,22 +1,19 @@
 import { useQuery } from '@tanstack/react-query';
 import { Trip } from '@/types/types'
 import { User } from "@auth/core/types";
+import { searchTrips } from "@/lib/utils/db-server-action";
 
 const fetchTrips = async (currentTrip: boolean): Promise<Array<Trip>> => {
   const response = await fetch(`/api/on-the-water?currentTrip=${currentTrip}`)
   return await response.json();
 }
-const searchTrips = async (guestName: string, departureTime: any): Promise<Array<Trip>> => {
-  const response = await fetch(`/api/search-trips?guestName=${guestName}&departureTime=${departureTime}`)
-  return await response.json();
-}
+
 export const useGetTrips = (currentTrip: boolean) => {
   return useQuery({
     queryKey: ['trips', currentTrip],
     queryFn: () => fetchTrips(currentTrip),
   })
 }
-
 
 const fetchUsers = async (): Promise<Array<User>> => {
   const response = await fetch(`/api/emails`)
@@ -33,6 +30,6 @@ export const useGetUser = () => {
 export const useGetSearchPageTrips = () => {
   return useQuery({
     queryKey: ['searchPageTrips'],
-    queryFn: () => searchTrips("", ""),
+    queryFn: () => searchTrips("", null),
   })
 }
